Relative Locations, Structures.
And the woes of unchecked compilation errors. Relatives are basically vectors, with a bit of a twist. With the structure classes the Multi-Block Structures don't need to store the slave list in NBT-data. Now it can just be part of the base definition and the slave list can be calculated from that. More cycles, less memory.